+++ lib/karafka/routing/builder.rb
@@ -32,6 +32,13 @@ module Karafka
         end
       end
 
+      # @return [Array<Karafka::Routing::ConsumerGroup>] only active consumer groups that
+      #   we want to use. Since Karafka supports multi-process setup, we need to be able
+      #   to pick only those consumer groups that should be active in our given process context
+      def active
+        select(&:active?)
+      end
+
       private
 
       # Builds and saves given consumer group