+++ spec/lib/karafka/routing/consumer_group_spec.rb
@@ -1,9 +1,9 @@
 # frozen_string_literal: true
 
 RSpec.describe Karafka::Routing::ConsumerGroup do
-  subject(:consumer_group) { described_class.new(id) }
+  subject(:consumer_group) { described_class.new(name) }
 
-  let(:id) { rand.to_s }
+  let(:name) { rand.to_s }
 
   describe 'after initialize' do
     it 'expect not to have any topics yet' do
@@ -13,7 +13,7 @@ RSpec.describe Karafka::Routing::ConsumerGroup do
 
   describe '#id' do
     it 'expect to namespace id with application client_id' do
-      expect(consumer_group.id).to eq "#{Karafka::App.config.client_id.to_s.underscore}_#{id}"
+      expect(consumer_group.id).to eq "#{Karafka::App.config.client_id.to_s.underscore}_#{name}"
     end
   end
 
@@ -52,6 +52,29 @@ RSpec.describe Karafka::Routing::ConsumerGroup do
 
     it { expect(consumer_group.topics.count).to eq 1 }
     it { expect(built_topic.name).to eq :topic_name.to_s }
+  end
+
+  describe '#active?' do
+    context 'when there are no topics in the consumer group' do
+      it { expect(consumer_group.active?).to eq false }
+    end
+
+    context 'when none of the topics is active' do
+      before do
+        consumer_group.public_send(:topic=, :topic_name) do
+          controller Class.new
+          worker Class.new
+        end
+      end
+
+      it { expect(consumer_group.active?).to eq false }
+    end
+
+    context 'when our consumer group name is in server consumer groups' do
+      before { Karafka::Server.consumer_groups = [name] }
+
+      it { expect(consumer_group.active?).to eq true }
+    end
   end
 
   describe '#to_h' do